requirements.txt
git+https://github.com/Rapptz/discord.py@master#egg=discord.py[voice]
git+https://github.com/Rapptz/discord-ext-menus
discord-flags >= 2.1.0
pyyaml >= 5.3
aiofiles >= 0.4.0
emoji
cachetools
psutil
matplotlib
